coverage: CR : convert `TestCover` convert back to `os/exec` pattern
